refactor: make loader as submodule, split to check/locate/load/cases
