+++ tests/test_response.py
@@ -9,7 +9,7 @@ from tests.base import ApiServerUnittest
 class TestResponse(ApiServerUnittest):
 
     def setUp(self):
-        self.functions_mapping = loader.load_module_functions(built_in)
+        self.functions_mapping = loader.load.load_module_functions(built_in)
 
     def test_parse_response_object_json(self):
         url = "http://127.0.0.1:5000/api/users"
