+++ tests/test_utils.py
@@ -63,7 +63,7 @@ class TestUtils(ApiServerUnittest):
 
     def current_validators(self):
         from httprunner import built_in
-        functions_mapping = loader.load_module_functions(built_in)
+        functions_mapping = loader.load.load_module_functions(built_in)
 
         functions_mapping["equals"](None, None)
         functions_mapping["equals"](1, 1)
